Moved ASTAR (class version) into task3.py, added control file using netowrkX
